+++ src/main/java/paymentStrategy/PaymentStrategy.java
@@ -3,5 +3,5 @@ package paymentStrategy;
 import model.PaymentStatus;
 
 public interface PaymentStrategy {
-    PaymentStatus makePayment(Long parkingFee);
+    PaymentStatus makePayment(Double parkingFee);
 }
