+++ src/main/java/paymentStrategy/UpiPaymentStrategy.java
@@ -4,7 +4,7 @@ import model.PaymentStatus;
 
 public class UpiPaymentStrategy implements PaymentStrategy {
     @Override
-    public PaymentStatus makePayment(Long parkingFee) {
+    public PaymentStatus makePayment(Double parkingFee) {
         System.out.println("Payment of Rs. " + parkingFee + " done through UPI");
         return PaymentStatus.SUCCESS;
     }
